requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cloudinary==1.44.1
Django==5.2.7
django-admin-rangefilter==0.13.3
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
